BC-1513 Move lesson controller out of index file
